- fixed bugs with flows - allow all nodes except to see all products - still have bug dealing with having a double price for BuyProductFlow param

+++ workflows/src/main/kotlin/com/r3/developers/csdetemplate/digitalcurrency/workflows/SellProductSubFlow.kt
@@ -118,7 +118,7 @@ class FinalizeSellProductResponderSubFlow: AbstractFlow(), ResponderFlow {
 
             val availableTokens = ledgerService.findUnconsumedStatesByType(DigitalCurrency::class.java)
             val coinSelection = CoinSelection()
-            val (currencyToSpend, spentCurrency) = coinSelection.selectTokensForTransfer(price.toInt(),
+            val (currencyToSpend, spentCurrency) = coinSelection.selectTokensForTransfer(price.toDouble(),
                 buyer.ledgerKeys.first(),
                 seller.ledgerKeys.first(),
                 availableTokens)
